Fix S3 tests
